ui/requirements.txt
flask>=2.3
Pillow>=9.0
